+++ cmd/healthcheck/healthcheck.go
@@ -7,10 +7,12 @@ import (
 	"github.com/ghodss/yaml"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	kerrors "k8s.io/apimachinery/pkg/util/errors"
 
 	"github.com/openshift/openshift-azure/pkg/api"
 	acsapi "github.com/openshift/openshift-azure/pkg/api"
 	"github.com/openshift/openshift-azure/pkg/plugin"
+	"github.com/openshift/openshift-azure/pkg/validate"
 )
 
 // healthCheck should get rolled into the end of createorupdate once the sync
@@ -26,6 +28,10 @@ func healthCheck() error {
 	err = yaml.Unmarshal(b, &cs)
 	if err != nil {
 		return errors.Wrap(err, "cannot unmarshal _data/containerservice.yaml")
+	}
+
+	if errs := validate.ContainerService(cs, nil); len(errs) > 0 {
+		return errors.Wrap(kerrors.NewAggregate(errs), "cannot validate _data/containerservice.yaml")
 	}
 
 	return p.HealthCheck(context.Background(), cs)
